pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repopulse"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "pydantic",
    "httpx",
    "pyyaml",
]

[project.scripts]
repopulse = "app.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["configs*", "tests*"]
